% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Laureen.R
\name{get_GUI}
\alias{get_GUI}
\title{get_GUI}
\usage{
get_GUI()
}
\arguments{
\item{none}{no parameters needed}
}
\value{
Returns a GUI of a Shiny App that displays rocket launches on a world map. These can be filtered according to success, provider & time frame. Furthermore, there is a statistics tab with some plots.
}
\description{
Function to Call the Shiny App "Rocket Launches 2024".
}
\note{
Error message "curl::curl_fetch_memory(url, handle = handle) : Timeout was reached: etc." might occur. Don't stop function in that case. The function will repeat itself automatically until GUI is returned.
}
\examples{
# get_GUI()
}
