% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Laureen.R
\name{get_GUI}
\alias{get_GUI}
\title{get_GUI}
\usage{
get_GUI()
}
\arguments{
\item{none}{no parameters needed}
}
\value{
Returns a GUI of a Shiny App that displays rocket launches on a world map. These can be filtered according to success, provider & time frame.
}
\description{
Function to Call the Shiny App "Rocket Launches 2024".
}
\note{
If Error occurs: "Error in curl::curl_fetch_memory(url, handle = handle) : etc", run get_GUI() again until the GUI is returned.
}
\examples{
# get_GUI()
}
